requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
cryptography==2.9.2
distlib==0.3.1
Django==3.0
django-filter==2.3.0
django-phonenumber-field==5.0.0
djangorestframework==3.10.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
Markdown==3.2.2
mysqlclient==2.0.1
phonenumbers==8.12.8
Pillow==7.2.0
pipenv==2020.8.13
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
